requirements.txt
aiohttp == 3.8.4
pymongo == 4.3.3
pyTelegramBotAPI == 4.10.0
python-dotenv == 1.0.0
requests == 2.28.2
urllib3 == 1.26.15
Flask == 2.2.3
pyngrok == 6.0.0
waitress == 2.1.2
pycodestyle == 2.10.0
